init: Let a device's init.rc redefine service entries
In case of duplicate service definitions, init.rc's version always
wins. This is a less-than-ideal situation and the main reason for
some devices to override the default init.rc, which leads to out-of-sync
situations whenever we happen to insert something into the main copy.

So let services be overridden. To prevent accidental override of
system services, use a specific keyword for this, "service_redefine"

Change-Id: I94eb4ebadfda8982128ec17f639411c03b18f68f
Signed-off-by: Louis Popi <theh2o64@gmail.com>